MeanDeviation: separate mean deviation from graphs
