+++ manifest.php
@@ -63,6 +63,26 @@ $actionRows[] = [
 
 //Exam Analysis: Mean score deviation per subject for selected exam type
 $actionRows[] = [
+    'name'                      => 'Mean Deviation Graphs', // The name of the action (appears to user in the right hand side module menu)
+    'precedence'                => '0',// If it is a grouped action, the precedence controls which is highest action in group
+    'category'                  => 'Analysis', // Optional: subgroups for the right hand side module menu
+    'description'               => 'This module is for administrators to do exam analysis', // Text description
+    'URLList'                   => 'meanDeviationGraphs.php', // List of pages included in this action
+    'entryURL'                  => 'meanDeviationGraphs.php', // The landing action for the page.
+    'entrySidebar'              => 'Y', // Whether or not there's a sidebar on entry to the action
+    'menuShow'                  => 'Y', // Whether or not this action shows up in menus or if it's hidden
+    'defaultPermissionAdmin'    => 'Y', // Default permission for built-in role Admin
+    'defaultPermissionTeacher'  => 'Y', // Default permission for built-in role Teacher
+    'defaultPermissionStudent'  => 'N', // Default permission for built-in role Student
+    'defaultPermissionParent'   => 'N', // Default permission for built-in role Parent
+    'defaultPermissionSupport'  => 'Y', // Default permission for built-in role Support
+    'categoryPermissionStaff'   => 'Y', // Should this action be available to user roles in the Staff category?
+    'categoryPermissionStudent' => 'N', // Should this action be available to user roles in the Student category?
+    'categoryPermissionParent'  => 'N', // Should this action be available to user roles in the Parent category?
+    'categoryPermissionOther'   => 'Y', // Should this action be available to user roles in the Other category?
+];
+//Exam Analysis: Mean score deviation per subject for selected exam type
+$actionRows[] = [
     'name'                      => 'Mean Score Deviation', // The name of the action (appears to user in the right hand side module menu)
     'precedence'                => '0',// If it is a grouped action, the precedence controls which is highest action in group
     'category'                  => 'Analysis', // Optional: subgroups for the right hand side module menu
